+++ src/main/java/org/server/remoteclass/controller/CouponController.java
@@ -1,11 +1,10 @@
 package org.server.remoteclass.controller;
 
 import io.swagger.annotations.ApiOperation;
-import org.server.remoteclass.dto.coupon.CouponDto;
 import org.server.remoteclass.dto.coupon.RequestCouponDto;
 import org.server.remoteclass.dto.coupon.ResponseCouponDto;
 import org.server.remoteclass.exception.IdNotExistException;
-import org.server.remoteclass.service.CouponService;
+import org.server.remoteclass.service.coupon.CouponService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
