+++ src/main/java/org/server/remoteclass/service/issuedCoupon/IssuedCouponServiceImpl.java
@@ -1,4 +1,4 @@
-package org.server.remoteclass.service;
+package org.server.remoteclass.service.issuedCoupon;
 
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
