+++ src/main/java/org/server/remoteclass/service/student/StudentService.java
@@ -1,4 +1,4 @@
-package org.server.remoteclass.service;
+package org.server.remoteclass.service.student;
 
 import org.server.remoteclass.dto.lecture.ResponseLectureDto;
 import org.server.remoteclass.dto.student.RequestStudentDto;
